Added multiple waypoint handling.
This required factoring out the waypoint stuff calculation so that it
could be called when the routine first initializes, then again as it
passes each waypoint.
