+++ ClosedLoopRobotController.py
@@ -1,4 +1,4 @@
-from Robot import RobotController
+from Robot import RobotController, coerceDHeadingRad
 
 class ClosedLoopRobotController(RobotController):
     def __init__(self,interface):
@@ -10,7 +10,7 @@ class ClosedLoopRobotController(RobotController):
         #Set up the steering control constants
         self.P=-20;
     def control(self):
-        stCmd=self.P*(self.heading-self.cmdHeading)
+        stCmd=self.P*coerceDHeadingRad(self.heading-self.cmdHeading)
         if stCmd>1:
             stCmd=1
         elif stCmd<-1:
